Add suspicious comment markings
In some cases there are reports of suspicious comments (e.g. a single
person writing multiple comments as if they were written by different
people, possibly because they hate somebody that much; or a professor
writing about themselves). Whenever it's possible to confirm it based on
the logs, it's nice to be able to mark comments as suspicious. This
commit implements that as a feature.